agents/iqvia_agent.py
import logging
from typing import Dict, Any, Optional
import requests
import os
from dotenv import load_dotenv

# Configure logging
logging.basicConfig(level=logging.INFO)
logger = logging.getLogger(__name__)

class IqvInsightsAgent:
    """Agent for fetching market insights from IQVIA API."""
    
    def __init__(self):
        load_dotenv()
        self.base_url = "https://api.iqvia.com"  # Replace with actual IQVIA API endpoint
        self.api_key = os.getenv('IQVIA_API_KEY')
        
    def get_market_insights(self, drug_name: str, therapeutic_area: str = None) -> Dict[str, Any]:
        """
        Get market insights for a specific drug.
        
        Args:
            drug_name: Name of the drug to analyze
            therapeutic_area: Optional therapeutic area to filter results
            
        Returns:
            Dict containing market insights data
        """
        try:
            # This is a placeholder implementation
            # In a real implementation, you would make API calls to IQVIA here
            
            # Mock response for demonstration
            return {
                'market_size': '1.2B USD',
                'cagr': '5.2%',
                'key_insights': [
                    f'Growing demand for {drug_name} in {therapeutic_area or "various therapeutic areas"}',
                    'Increasing market competition observed',
                    'Patent expiry expected in 2027'
                ],
                'data_source': 'IQVIA Market Intelligence',
                'last_updated': '2024-03-15'
            }
            
        except Exception as e:
            logger.error(f"Error fetching market insights: {str(e)}")
            return {
                'error': str(e),
                'market_size': 'N/A',
                'cagr': 'N/A',
                'key_insights': ['Failed to fetch market insights']
            }
